+++ packages/combobox/src/types.ts
@@ -105,6 +105,4 @@ export interface Options {
   autoselect?: boolean;
   write?: Write;
   render?: Render;
-  /** Optional open button (APG). */
-  button?: HTMLButtonElement | null;
 }